Workaround for flutter 3.10.
FixedScrollMetrics constructor now has a devicePixelRatio required
named parameter.

For the time being, we get the value from the first view in
PlatformDispatcher.instance.views.

This will work correctly on single-window applications (which are still
the only option), but may pose a problem in the future when multiple
view support is implemented in flutter.